API

ExistingProcessManagers.julia_worker_regex

ExistingProcessManagers.ExistingProcessManager

ExistingProcessManagers.ExistingProcessManager

ExistingProcessManagers.ExistingProcessManager

ExistingProcessManagers.addprocs_existing

ExistingProcessManagers.hosts_and_ports_to_workerconfigs

ExistingProcessManagers.parse_julia_worker_output_to_hosts_and_ports

ExistingProcessManagers.julia_worker_regex — Constant

julia_worker_regex

Regex for parsing the output printed by Julia worker processes.

ExistingProcessManagers.ExistingProcessManager — Type

struct ExistingProcessManager <: Distributed.ClusterManager

Fields:

wconfigs::Array{Distributed.WorkerConfig,1}

ExistingProcessManager(wconfigs::Vector{Distributed.WorkerConfig})

Construct an ExistingProcessManager from a list of WorkerConfigs.

Example

julia> w1 = WorkerConfig()
WorkerConfig(nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing)

julia> w1.host = "127.0.0.1"
"127.0.0.1"

julia> w1.port = 9601
9601

julia> w2 = WorkerConfig()
WorkerConfig(nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing)

julia> w2.host = "127.0.0.1"
"127.0.0.1"

julia> w2.port = 9602
9602

julia> wconfigs = WorkerConfig[w1, w2]
2-element Vector{WorkerConfig}:
 WorkerConfig(nothing, "127.0.0.1", 9601, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing)
 WorkerConfig(nothing, "127.0.0.1", 9602, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing)

julia> manager = ExistingProcessManager(wconfigs)
ExistingProcessManager(WorkerConfig[WorkerConfig(nothing, "127.0.0.1", 9601, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing), WorkerConfig(nothing, "127.0.0.1", 9602, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing)])

julia> addprocs(manager)

ExistingProcessManagers.ExistingProcessManager — Method

ExistingProcessManager(worker_output::AbstractString)

Construct an ExistingProcessManager from the output printed by Julia worker processes.

Example

julia> worker_output = """
       julia_worker:9960#192.168.1.151
       julia_worker:9960#192.168.1.151
       julia_worker:9960#192.168.1.151
       julia_worker:9966#192.168.1.157
       julia_worker:9968#192.168.1.159
       """
"julia_worker:9960#192.168.1.151
julia_worker:9960#192.168.1.151
julia_worker:9960#192.168.1.151
julia_worker:9966#192.168.1.157
julia_worker:9968#192.168.1.159
"

julia> manager = ExistingProcessManager(worker_output)
ExistingProcessManager(WorkerConfig[WorkerConfig(nothing, "192.168.1.151", 9960, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing), WorkerConfig(nothing, "192.168.1.157", 9966, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing), WorkerConfig(nothing, "192.168.1.159", 9968, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing)])

julia> addprocs(manager)

ExistingProcessManagers.ExistingProcessManager — Method

ExistingProcessManager(hosts_and_ports::Vector{Tuple{String, Int}})

Construct an ExistingProcessManager from a list of hosts and port numbers.

Example

julia> hosts_and_ports = [
       ("127.0.0.1", 9601), # the host is "127.0.0.1", and the port number is 9601
       ("127.0.0.1", 9602), # the host is "127.0.0.1", and the port number is 9602
       ]
2-element Vector{Tuple{String, Int64}}:
 ("127.0.0.1", 9601)
 ("127.0.0.1", 9602)

julia> manager = ExistingProcessManager(hosts_and_ports)
ExistingProcessManager(WorkerConfig[WorkerConfig(nothing, "127.0.0.1", 9601, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing), WorkerConfig(nothing, "127.0.0.1", 9602, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing)])

julia> addprocs(manager)

ExistingProcessManagers.addprocs_existing — Method

addprocs_existing(workers; kwargs...)

Equivalent to addprocs(ExistingProcessManager(workers); kwargs...).

Examples

Example 1

julia> hosts_and_ports = [
       ("127.0.0.1", 9601), # the host is "127.0.0.1", and the port number is 9601
       ("127.0.0.1", 9602), # the host is "127.0.0.1", and the port number is 9602
       ]
2-element Vector{Tuple{String, Int64}}:
 ("127.0.0.1", 9601)
 ("127.0.0.1", 9602)

julia> addprocs_existing(hosts_and_ports; kwargs...)

Example 2

julia> w1 = WorkerConfig()
WorkerConfig(nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing)

julia> w1.host = "127.0.0.1"
"127.0.0.1"

julia> w1.port = 9601
9601

julia> w2 = WorkerConfig()
WorkerConfig(nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing)

julia> w2.host = "127.0.0.1"
"127.0.0.1"

julia> w2.port = 9602
9602

julia> wconfigs = WorkerConfig[w1, w2]
2-element Vector{WorkerConfig}:
 WorkerConfig(nothing, "127.0.0.1", 9601, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing)
 WorkerConfig(nothing, "127.0.0.1", 9602, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing)

julia> addprocs_existing(wconfigs; kwargs...)

Example 3

julia> worker_output = """
       julia_worker:9960#192.168.1.151
       julia_worker:9960#192.168.1.151
       julia_worker:9960#192.168.1.151
       julia_worker:9966#192.168.1.157
       julia_worker:9968#192.168.1.159
       """
"julia_worker:9960#192.168.1.151
julia_worker:9960#192.168.1.151
julia_worker:9960#192.168.1.151
julia_worker:9966#192.168.1.157
julia_worker:9968#192.168.1.159
"

julia> addprocs_existing(worker_output; kwargs...)

ExistingProcessManagers.hosts_and_ports_to_workerconfigs — Method

hosts_and_ports_to_workerconfigs(hosts_and_ports::Vector{Tuple{String, Int}})

Convert a list of hosts and port numbers to a list of WorkerConfigs.

Examples

julia> hosts_and_ports = [
       ("127.0.0.1", 9601), # the host is "127.0.0.1", and the port number is 9601
       ("127.0.0.1", 9602), # the host is "127.0.0.1", and the port number is 9602
       ]
2-element Vector{Tuple{String, Int64}}:
 ("127.0.0.1", 9601)
 ("127.0.0.1", 9602)

julia> hosts_and_ports_to_workerconfigs(hosts_and_ports)
2-element Vector{Distributed.WorkerConfig}:
 Distributed.WorkerConfig(nothing, "127.0.0.1", 9601, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing)
 Distributed.WorkerConfig(nothing, "127.0.0.1", 9602, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing, nothing)

ExistingProcessManagers.parse_julia_worker_output_to_hosts_and_ports — Method

parse_julia_worker_output_to_hosts_and_ports(; regex = julia_worker_regex)

Parse the output printed by Julia workers and return a list of hosts and ports.

Example

julia> worker_output = """
       julia_worker:9960#192.168.1.151
       julia_worker:9962#192.168.1.153
       julia_worker:9964#192.168.1.155
       julia_worker:9966#192.168.1.157
       julia_worker:9968#192.168.1.159
       """
"julia_worker:9960#192.168.1.151
julia_worker:9962#192.168.1.153
julia_worker:9964#192.168.1.155
julia_worker:9966#192.168.1.157
julia_worker:9968#192.168.1.159
"

julia> parse_julia_worker_output_to_hosts_and_ports(worker_output)
5-element Vector{Tuple{String, Int64}}:
 ("192.168.1.151", 9960)
 ("192.168.1.153", 9962)
 ("192.168.1.155", 9964)
 ("192.168.1.157", 9966)
 ("192.168.1.159", 9968)